libpod: ensure console socket buffers are properly sized
On FreeBSD, the default socket buffer size for unix domain sockets is
just 8192 which is too small for workloads which transmit large amounts
of data into the container via stdin. This caused the 'podman exec - cat
from stdin' system test to hang when writing large amounts of data into
a container.  To avoid the problem, this commit increases the buffer
sizes to 32768 which is enough to allow the test to pass reliably.

Signed-off-by: Doug Rabson <dfr@rabson.org>